Drop read_config_file, refactor child_register
Drop function `read_config_file`. It does nothing more than return a tuple of
values from the config file. The contents of the config file are available to
all functions via the global variable `CONFIG`, so this extra level of
indirection is not necessary.

Refactor function `child_register`. It made the most extensive use of
`read_config_file` via this line of code:

    adminpassword, orgname, contentview = read_config_file()[4:7]

It was also a bit messy in general.

Both the Python 2 and 3 versions of flake8 and pylint are happy with the script.