Fix config file parsing
Add `configparser` to the list of requirements. ConfigParser saw a major update
in Python 3.2, and this module backports those changes to Python 2.6.  The
default configparser module is used in Python 3.2+, and the backported version
is used in older versions of Python.

As a side-effect of this change, drop a conditional import and do not bother
inspecting `sys.version_info`.